cloud-storage-deploy/requirements.txt
Flask==2.3.3
Flask-Session==0.5.0
mysql-connector-python==8.1.0
boto3==1.28.57
Werkzeug==2.3.7
python-dotenv==1.0.0
botocore==1.31.57
